+++ ebegu-testutils/src/main/java/ch/dvbern/ebegu/test/TestDataUtil.java
@@ -565,7 +565,7 @@ public final class TestDataUtil {
 
 	public static PensumFachstelle createDefaultPensumFachstelle() {
 		PensumFachstelle pensumFachstelle = new PensumFachstelle();
-		pensumFachstelle.setPensum(50l);
+		pensumFachstelle.setPensum(50);
 		pensumFachstelle.setGueltigkeit(new DateRange(LocalDate.now(), Constants.END_OF_TIME));
 		pensumFachstelle.setFachstelle(createDefaultFachstelle());
 		return pensumFachstelle;
@@ -602,7 +602,7 @@ public final class TestDataUtil {
 		return epCont;
 	}
 
-	public static ErwerbspensumContainer createErwerbspensum(LocalDate von, LocalDate bis, long pensum, int zuschlag) {
+	public static ErwerbspensumContainer createErwerbspensum(LocalDate von, LocalDate bis, int pensum, int zuschlag) {
 		ErwerbspensumContainer erwerbspensumContainer = new ErwerbspensumContainer();
 		Erwerbspensum erwerbspensum = new Erwerbspensum();
 		erwerbspensum.setPensum(pensum);
@@ -615,7 +615,7 @@ public final class TestDataUtil {
 	public static Erwerbspensum createErwerbspensumData() {
 		Erwerbspensum ep = new Erwerbspensum();
 		ep.setTaetigkeit(Taetigkeit.ANGESTELLT);
-		ep.setPensum(50l);
+		ep.setPensum(50);
 		ep.setZuschlagZuErwerbspensum(true);
 		ep.setZuschlagsgrund(Zuschlagsgrund.LANGER_ARBWEITSWEG);
 		ep.setZuschlagsprozent(10);
@@ -659,7 +659,7 @@ public final class TestDataUtil {
 
 	private static Betreuungspensum createBetreuungspensum() {
 		Betreuungspensum betreuungspensum = new Betreuungspensum();
-		betreuungspensum.setPensum(80l);
+		betreuungspensum.setPensum(80);
 		return betreuungspensum;
 	}
 
@@ -1313,7 +1313,7 @@ public final class TestDataUtil {
 		BetreuungsmitteilungPensum pensum = new BetreuungsmitteilungPensum();
 		pensum.setBetreuungsmitteilung(mitteilung);
 		pensum.setGueltigkeit(new DateRange(Constants.START_OF_TIME, Constants.END_OF_TIME));
-		pensum.setPensum(30l);
+		pensum.setPensum(30);
 
 		betPensen.add(pensum);
 		mitteilung.setBetreuungspensen(betPensen);
